requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic-settings==2.2.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
redis==5.0.4
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.1
alembic==1.13.1
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.6
flake8==7.0.0
bandit==1.7.8